devices: Refactor hotplugVFIODevice() to have common exit path
hotplugVFIODevice() has several different paths depending if we're plugging
into a root port or a PCIE<->PCI bridge and if we're using a regular
or mediated VFIO device.

We're going to want some common code on the successful exit path here, so
refactor the function to allow that without duplication.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>