device: Tell agent to rebind VFIO devices to guest VFIO driver
We send information about several kinds of devices to the agent so that it
can apply specific handling.  We don't currently do this with VFIO devices,
but we will need to in order to improve handling of SR-IOV devices.

This patch sends information about VFIO devices to the agent, if the
vfio_in_guest configuration option is enabled.  Corresponding agent code
will use that to properly set up VFIO devices in the guest.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>