device: Give the agent information about VFIO devices
We send information about several kinds of devices to the agent so that it
can apply specific handling.  We don't currently do this with VFIO devices.
However we need to do that so that the agent can properly wait for VFIO
devices to be ready (previously it did that using a PCI rescan which may
not be reliable and has some very bad side effects).

This patch collates and sends the relevant information.

Depends-on: github.com/kata-containers/agent#850
fixes #2664

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>