devices: Record guest PCI path for VFIO devices
For several device types which correspond to a PCI device in the guest
we record the device's PCI address in the guest.  We don't currently do
that for VFIO devices, but we're going to need to for better handling of
SR-IOV devices.

Add code to do this, with a couple of wrinkles from existing examples:
 * We don't know the address the VFIO devices before we hotplug them in
   all cases, so we need to query the address back from qemu using qom-get.
 * We use the name 'GuestPciPath' instead of 'PCIAddr' to reduce confusion
   with more typical DDDD:BB:DD.F style PCI addresses.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>