device: Add vfio_in_guest option
On a conventional (e.g. runc) container, passing in a VFIO group device,
/dev/vfio/NN, will result in the same VFIO group device being available
within the container.

With Kata, however, the VFIO device will be bound to the guest kernel's
driver (if available), possibly appearing as some interpreted device within
the guest.

This adds a new option `vfio_in_guest` which makes this behaviour
difference explicit.  If set to false, it keeps the current Kata
behaviour of binding VFIO devices to the guest native driver.  If set
to true it will remap VFIO devices to the VFIO driver within the guest
as well, meaning they will appear as VFIO devices within the guest.

Unlike a runc container, the VFIO devices will have different names to the
host, since the names correspond to the IOMMU groups of the guest and those
can't be remapped with namespaces.

For now we keep 'false' as the default value, to maintain current Kata
behaviour.  In future we should change this to 'true', so that Kata's
default behaviour more closely resembles OCI specified behaviour.

fixes #2938

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>